# TurbineOps-Lite-Starter/backend/src/rest/routes/finding.routes.ts
import { Router } from 'express';
import { asyncHandler } from '../../utils/asyncHandlers';
import { searchFindings, updateFinding } from '../../services/finding.service';
import { canWrite } from '../../middleware/auth';

export const findingRoutes = Router();

findingRoutes.patch('/:id', canWrite, asyncHandler(async (req, res) => {
  res.json(await updateFinding(req.params.id, req.body || {}));
}));

findingRoutes.get('/', asyncHandler(async (req, res) => {
  res.json(await searchFindings(req.query.q));
}));